+++ src/biodwh2-drugcentral/src/main/java/de.unibi.agbi.biodwh2.drugcentral/model/TargetComponent.java
@@ -2,12 +2,13 @@ package de.unibi.agbi.biodwh2.drugcentral.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import de.unibi.agbi.biodwh2.core.model.graph.GraphProperty;
 import de.unibi.agbi.biodwh2.core.model.graph.GraphNodeLabel;
+import de.unibi.agbi.biodwh2.core.model.graph.GraphProperty;
+import de.unibi.agbi.biodwh2.drugcentral.etl.DrugCentralGraphExporter;
 
 @SuppressWarnings("unused")
 @JsonPropertyOrder({"id", "accession", "swissprot", "organism", "name", "gene", "geneid", "tdl"})
-@GraphNodeLabel("TargetComponent")
+@GraphNodeLabel(DrugCentralGraphExporter.TARGET_COMPONENT_LABEL)
 public final class TargetComponent {
     @JsonProperty("id")
     public Integer id;